+++ ngClient/App/settings/Setup/googleAnalytics.js
@@ -1,0 +1,8 @@
+﻿(function () {
+    'use strict';
+
+    angular.module('main')
+        .constant('AnalyticsTrackingCode', '') // Enter your google analytics tracking code here
+        .constant('AnalyticsDomainName', 'none'); // Enter your domain name here - 'none' for localhost
+
+})();