+++ ngClient/App/settings/Setup/serviceAppUrl.js
@@ -1,7 +1,7 @@
 ﻿(function () {
     'use strict';
 
-    angular.module('main')
-        .constant('serviceAppUrl', 'http://localhost:15001');
+    // Enter the url of service application (WebApi) here
+    angular.module('main').constant('serviceAppUrl', 'http://localhost:15001');
 
 })();