+++ SolutionItems/Documents/Release Notes.md
@@ -3,6 +3,7 @@
 **0.43**
 
 * gulp: auto jshint check, bundling, minifying for js & css files
+* google analytics is now handled through "angular-google-analytics"
 * jshint fixes
 * ResourcePoolEdit(Controller) was renamed to ResourcePoolManage(Controller) to fix the conflict with generated one